requirements.txt
asyncpg==0.30.0
fastapi==0.115.12
opencv-python==4.11.0.86
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
python-telegram-bot==22.0
uvicorn==0.34.1
